Refactor to extract FrostedGlassColumn component
